moved AdminEnvironment to seperate file
fixes: #48038
