+++ js/apps/admin-ui/src/index.ts
@@ -322,7 +322,7 @@ export { Header } from "./PageHeader";
 export { PageNav } from "./PageNav";
 export { PageNotFoundSection } from "./PageNotFoundSection";
 export { App as AdminUi } from "./App";
-export type { Environment as AdminEnvironment } from "./environment";
+export type { Environment as AdminEnvironment } from "./environment.d";
 export { KeycloakProvider, useEnvironment } from "@keycloak/keycloak-ui-shared";
 export { AdminClientContext, initAdminClient } from "./admin-client";
 export { AppContexts } from "./App";
