+++ js/apps/admin-ui/src/routes.tsx
@@ -1,9 +1,8 @@
 import type { AccessType } from "@keycloak/keycloak-admin-client/lib/defs/whoAmIRepresentation";
 import type { TFunction } from "i18next";
 import type { ComponentType } from "react";
-import type { NonIndexRouteObject, RouteObject } from "react-router-dom";
+import type { NonIndexRouteObject } from "react-router-dom";
 import { PageNotFoundSection } from "./PageNotFoundSection";
-import { Root } from "./Root";
 import authenticationRoutes from "./authentication/routes";
 import clientScopesRoutes from "./client-scopes/routes";
 import clientRoutes from "./clients/routes";
@@ -60,9 +59,3 @@ export const routes: AppRouteObject[] = [
   ...pageRoutes,
   NotFoundRoute,
 ];
-
-export const RootRoute: RouteObject = {
-  path: "/",
-  element: <Root />,
-  children: routes,
-};
